# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Top Maracana API",
        "description": "Documentation de l'API Top Maracana",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/teams": {
            "get": {
                "tags": [
                    "Équipes"
                ],
                "summary": "Lister toutes les équipes",
                "operationId": "f44e2fcbf1cbf47247272fa5e2632ecb",
                "responses": {
                    "200": {
                        "description": "Liste des équipes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "teams": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "string",
                                                        "example": "550e8400-e29b-41d4-a716-446655440000"
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Les Lions"
                                                    },
                                                    "logo": {
                                                        "type": "string",
                                                        "example": "teams/logos/logo.png"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Équipes"
                ],
                "summary": "Créer une équipe et s'inscrire à une compétition",
                "operationId": "b86cd7af54a3695d92ddfbd0c2747e71",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "competition_id",
                                    "name",
                                    "players"
                                ],
                                "properties": {
                                    "competition_id": {
                                        "type": "string",
                                        "example": "550e8400-e29b-41d4-a716-446655440000"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Les Lions"
                                    },
                                    "players": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "full_name": {
                                                    "type": "string",
                                                    "example": "Dimitri Ouedraogo"
                                                },
                                                "birth_date": {
                                                    "type": "string",
                                                    "example": "2000-01-01"
                                                },
                                                "national_id_number": {
                                                    "type": "string",
                                                    "example": "BF123456"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Équipe créée et inscrite avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "team": {
                                            "properties": {
                                                "id": {
                                                    "type": "string",
                                                    "example": "550e8400-e29b-41d4-a716-446655440000"
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Les Lions"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erreur de validation métier"
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/teams/my": {
            "get": {
                "tags": [
                    "Équipes"
                ],
                "summary": "Mes équipes",
                "operationId": "32a03aa6e4439237b66a392eb9fc81b7",
                "responses": {
                    "200": {
                        "description": "Mes équipes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "teams": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "string",
                                                        "example": "550e8400-e29b-41d4-a716-446655440000"
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Les Lions"
                                                    },
                                                    "competition": {
                                                        "properties": {
                                                            "name": {
                                                                "type": "string",
                                                                "example": "Tournoi Maracana"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/teams/{id}": {
            "get": {
                "tags": [
                    "Équipes"
                ],
                "summary": "Détail d'une équipe",
                "operationId": "5f7cec6639feb6f6205ee6aeb30c1682",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "UUID de l'équipe",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "550e8400-e29b-41d4-a716-446655440000"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détail de l'équipe",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "team": {
                                            "properties": {
                                                "id": {
                                                    "type": "string",
                                                    "example": "550e8400-e29b-41d4-a716-446655440000"
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Les Lions"
                                                },
                                                "players": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "full_name": {
                                                                "type": "string",
                                                                "example": "Dimitri Ouedraogo"
                                                            },
                                                            "birth_date": {
                                                                "type": "string",
                                                                "example": "2000-01-01"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Équipe introuvable"
                    }
                }
            },
            "post": {
                "tags": [
                    "Équipes"
                ],
                "summary": "Modifier une équipe",
                "operationId": "1848fdd95dbe8d3acfdf43f6242c0b0a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "UUID de l'équipe",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "550e8400-e29b-41d4-a716-446655440000"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Les Lions FC"
                                    },
                                    "players": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "full_name": {
                                                    "type": "string",
                                                    "example": "Dimitri Ouedraogo"
                                                },
                                                "birth_date": {
                                                    "type": "string",
                                                    "example": "2000-01-01"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Équipe modifiée avec succès"
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Action non autorisée"
                    },
                    "404": {
                        "description": "Équipe introuvable"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Équipes"
                ],
                "summary": "Supprimer une équipe",
                "operationId": "1d3e81060958eaef03b74307a982b913",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "UUID de l'équipe",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "550e8400-e29b-41d4-a716-446655440000"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Équipe supprimée avec succès"
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Action non autorisée"
                    },
                    "404": {
                        "description": "Équipe introuvable"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}